@@ -3,28 +3,27 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.5",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
     "@mui/lab": "^5.0.0-alpha.62",
-    "@mui/material": "^5.2.5",
-    "@mui/styles": "^5.3.0",
-    "devextreme": "21.2.4",
-    "devextreme-react": "21.2.4",
-    "firebase": "^8.10.1",
-    "moment": "^2.29.2",
+    "@mui/material": "^5.11.4",
+    "devextreme": "22.2.3",
+    "devextreme-react": "22.2.3",
+    "firebase": "^9.15.0",
+    "moment": "^2.29.4",
     "moment-range": "^4.0.2",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-csv": "^2.2.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.2.2",
-    "replace-in-file": "^6.3.2",
-    "source-map-explorer": "^2.5.2",
-    "uuid": "^8.3.2",
-    "yargs": "^17.3.1"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.7.0",
+    "replace-in-file": "^6.3.5",
+    "source-map-explorer": "^2.5.3",
+    "uuid": "^9.0.0",
+    "yargs": "^17.6.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -57,10 +56,10 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.5.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "prettier": "^2.5.1"
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.0",
+    "prettier": "^2.8.2"
   }
 }
